Fix disappearing 3D molecule after new upload or delete It seems like the <canvas> element disappears after an upload or a delete. This fix uses the auto-initialization function of 3Dmol.js (https://github.com/3dmol/3Dmol.js/blob/master/3Dmol/autoload.js) which is running after a webpage is loaded ($(document).ready(function()) if 3Dmol.js is included and generates the <canvas> element. We have to avoid that it runs a second time and generates a second <canvas> element, run it after AJAX is done.
